LPS-75989 Move UserFormConstants to users-admin-api.
