+++ modules/apps/foundation/users-admin/users-admin-web/src/main/resources/META-INF/resources/init.jsp
@@ -176,12 +176,12 @@ page import="com.liferay.roles.admin.kernel.util.RolesAdminUtil" %><%@
 page import="com.liferay.taglib.search.ResultRow" %><%@
 page import="com.liferay.taglib.search.SearchEntry" %><%@
 page import="com.liferay.users.admin.configuration.UserFileUploadsConfiguration" %><%@
+page import="com.liferay.users.admin.constants.UserFormConstants" %><%@
 page import="com.liferay.users.admin.constants.UsersAdminPortletKeys" %><%@
 page import="com.liferay.users.admin.display.context.InitDisplayContext" %><%@
 page import="com.liferay.users.admin.display.context.UserDisplayContext" %><%@
 page import="com.liferay.users.admin.kernel.util.UsersAdmin" %><%@
 page import="com.liferay.users.admin.kernel.util.UsersAdminUtil" %><%@
-page import="com.liferay.users.admin.web.constants.UserFormConstants" %><%@
 page import="com.liferay.users.admin.web.constants.UsersAdminWebKeys" %><%@
 page import="com.liferay.users.admin.web.search.AddUserOrganizationChecker" %><%@
 page import="com.liferay.users.admin.web.search.OrganizationChecker" %><%@
